Fixed problem with Image.md
The failure inducing input is Image.md.

The symptom is that running the code on Image.md includes the image in output:
[hi.jpg]

The bug is that the code recognizes an image as a link- we fixed this by only adding the "link" (according to the previous version of the code) if it is not actually an image.